Backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
click==8.3.1
colorama==0.4.6
exceptiongroup==1.3.1
fastapi==0.135.3
filelock==3.25.2
fsspec==2026.3.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.8.0
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
packaging==26.0
pandas==2.3.3
pillow==12.2.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
python-dateutil==2.9.0.post0
python-multipart==0.0.22
pytz==2026.1.post1
PyYAML==6.0.3
regex==2026.3.32
rich==14.3.3
safetensors==0.7.0
shellingham==1.5.4
six==1.17.0
starlette==1.0.0
sympy==1.14.0
tokenizers==0.22.2
torch==2.11.0
torchvision==0.26.0
tqdm==4.67.3
transformers==5.5.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.42.0
